+++ src/Microsoft.Health.Dicom.Client/DicomWebClient.cs
@@ -17,7 +17,6 @@ using Dicom;
 using Dicom.Serialization;
 using EnsureThat;
 using Microsoft.AspNetCore.WebUtilities;
-using Microsoft.Health.Dicom.Client.Models;
 using Microsoft.Net.Http.Headers;
 using Newtonsoft.Json;
 using MediaTypeHeaderValue = Microsoft.Net.Http.Headers.MediaTypeHeaderValue;
@@ -57,619 +56,7 @@ namespace Microsoft.Health.Dicom.Client
         /// </remarks>
         public Func<MemoryStream> GetMemoryStream { get; set; }
 
-        #region WADO-RS
-        public async Task<DicomWebAsyncEnumerableResponse<DicomFile>> RetrieveStudyAsync(
-            string studyInstanceUid,
-            string dicomTransferSyntax = DicomWebConstants.OriginalDicomTransferSyntax,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
 
-            return await RetrieveInstancesAsync(
-                new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseStudyUriFormat, studyInstanceUid), UriKind.Relative),
-                dicomTransferSyntax,
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> RetrieveStudyMetadataAsync(
-            string studyInstanceUid,
-            string ifNoneMatch = null,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-
-            return await RetrieveMetadataAsync(
-                new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseRetrieveStudyMetadataUriFormat, studyInstanceUid), UriKind.Relative),
-                ifNoneMatch,
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomFile>> RetrieveSeriesAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string dicomTransferSyntax = DicomWebConstants.OriginalDicomTransferSyntax,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-
-            return await RetrieveInstancesAsync(
-                new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseSeriesUriFormat, studyInstanceUid, seriesInstanceUid), UriKind.Relative),
-                dicomTransferSyntax,
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> RetrieveSeriesMetadataAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string ifNoneMatch = null,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-
-            return await RetrieveMetadataAsync(
-                new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseRetrieveSeriesMetadataUriFormat, studyInstanceUid, seriesInstanceUid), UriKind.Relative),
-                ifNoneMatch,
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebResponse<DicomFile>> RetrieveInstanceAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string sopInstanceUid,
-            string dicomTransferSyntax = DicomWebConstants.OriginalDicomTransferSyntax,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(sopInstanceUid, nameof(sopInstanceUid));
-
-            return await RetrieveInstanceAsync(
-                new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseInstanceUriFormat, studyInstanceUid, seriesInstanceUid, sopInstanceUid), UriKind.Relative),
-                dicomTransferSyntax,
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> RetrieveInstanceMetadataAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string sopInstanceUid,
-            string ifNoneMatch = null,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(sopInstanceUid, nameof(sopInstanceUid));
-
-            return await RetrieveMetadataAsync(
-                new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseRetrieveInstanceMetadataUriFormat, studyInstanceUid, seriesInstanceUid, sopInstanceUid), UriKind.Relative),
-                ifNoneMatch,
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<Stream>> RetrieveFramesAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string sopInstanceUid,
-            int[] frames = null,
-            string mediaType = DicomWebConstants.ApplicationOctetStreamMediaType,
-            string dicomTransferSyntax = DicomWebConstants.OriginalDicomTransferSyntax,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(sopInstanceUid, nameof(sopInstanceUid));
-            var requestUri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseRetrieveFramesUriFormat, studyInstanceUid, seriesInstanceUid, sopInstanceUid, string.Join("%2C", frames)), UriKind.Relative);
-
-            return await RetrieveFramesAsync(requestUri, mediaType, dicomTransferSyntax, cancellationToken).ConfigureAwait(false);
-        }
-
-        private async Task<DicomWebResponse<DicomFile>> RetrieveInstanceAsync(
-            Uri requestUri,
-            string dicomTransferSyntax,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(requestUri, nameof(requestUri));
-
-            using var request = new HttpRequestMessage(HttpMethod.Get, requestUri);
-
-            request.Headers.TryAddWithoutValidation(
-                "Accept",
-                CreateAcceptHeader(DicomWebConstants.MediaTypeApplicationDicom, dicomTransferSyntax));
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebResponse<DicomFile>(
-                response,
-                async content =>
-                {
-                    MemoryStream memoryStream = GetMemoryStream();
-                    await content.CopyToAsync(memoryStream).ConfigureAwait(false);
-                    memoryStream.Seek(0, SeekOrigin.Begin);
-
-                    return await DicomFile.OpenAsync(memoryStream).ConfigureAwait(false);
-                });
-        }
-
-        private async Task<DicomWebAsyncEnumerableResponse<DicomFile>> RetrieveInstancesAsync(
-            Uri requestUri,
-            string dicomTransferSyntax,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(requestUri, nameof(requestUri));
-
-            using var request = new HttpRequestMessage(HttpMethod.Get, requestUri);
-
-            request.Headers.TryAddWithoutValidation(
-                "Accept",
-                CreateAcceptHeader(CreateMultipartMediaTypeHeader(DicomWebConstants.ApplicationDicomMediaType), dicomTransferSyntax));
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebAsyncEnumerableResponse<DicomFile>(
-                response,
-                ReadMultipartResponseAsDicomFileAsync(response.Content, cancellationToken));
-        }
-
-        private async Task<DicomWebAsyncEnumerableResponse<Stream>> RetrieveFramesAsync(
-            Uri requestUri,
-            string mediaType,
-            string dicomTransferSyntax,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(requestUri, nameof(requestUri));
-            EnsureArg.IsNotNullOrWhiteSpace(mediaType, nameof(mediaType));
-
-            using var request = new HttpRequestMessage(HttpMethod.Get, requestUri);
-
-            request.Headers.TryAddWithoutValidation(
-                "Accept",
-                CreateAcceptHeader(CreateMultipartMediaTypeHeader(mediaType), dicomTransferSyntax));
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebAsyncEnumerableResponse<Stream>(
-                response,
-                ReadMultipartResponseAsStreamsAsync(response.Content, cancellationToken));
-        }
-
-        private async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> RetrieveMetadataAsync(
-            Uri requestUri,
-            string ifNoneMatch,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(requestUri, nameof(requestUri));
-
-            using var request = new HttpRequestMessage(HttpMethod.Get, requestUri);
-
-            request.Headers.Accept.Add(DicomWebConstants.MediaTypeApplicationDicomJson);
-
-            if (!string.IsNullOrEmpty(ifNoneMatch))
-            {
-                request.Headers.TryAddWithoutValidation(HeaderNames.IfNoneMatch, ifNoneMatch);
-            }
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(
-                response,
-                additionalFailureInspector: (statusCode, responseHeaders, contentHeaders, responseBody) =>
-                {
-                    // If the content has not changed, the status returned will be NotModified and so we need to treat it specially.
-                    return statusCode == HttpStatusCode.NotModified;
-                })
-                .ConfigureAwait(false);
-
-            return new DicomWebAsyncEnumerableResponse<DicomDataset>(
-                response,
-                DeserializeAsAsyncEnumerable<DicomDataset>(response.Content));
-        }
-        #endregion
-
-        #region STOW-RS
-        public async Task<DicomWebResponse<DicomDataset>> StoreAsync(
-            IEnumerable<DicomFile> dicomFiles,
-            string studyInstanceUid,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(dicomFiles, nameof(dicomFiles));
-
-            var postContent = new List<Stream>();
-
-            try
-            {
-                foreach (DicomFile dicomFile in dicomFiles)
-                {
-                    MemoryStream stream = GetMemoryStream();
-                    await dicomFile.SaveAsync(stream).ConfigureAwait(false);
-                    stream.Seek(0, SeekOrigin.Begin);
-                    postContent.Add(stream);
-                }
-
-                using MultipartContent content = ConvertStreamsToMultipartContent(postContent);
-                return await StoreAsync(
-                    GenerateStoreRequestUri(studyInstanceUid),
-                    content,
-                    cancellationToken).ConfigureAwait(false);
-            }
-            finally
-            {
-                foreach (Stream stream in postContent)
-                {
-                    await stream.DisposeAsync().ConfigureAwait(false);
-                }
-            }
-        }
-
-        public async Task<DicomWebResponse<DicomDataset>> StoreAsync(
-            IEnumerable<Stream> streams,
-            string studyInstanceUid,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(streams, nameof(streams));
-
-            return await StoreAsync(
-                GenerateStoreRequestUri(studyInstanceUid),
-                ConvertStreamsToMultipartContent(streams),
-                cancellationToken)
-                .ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebResponse<DicomDataset>> StoreAsync(
-            Stream stream,
-            string studyInstanceUid,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(stream, nameof(stream));
-
-            stream.Seek(0, SeekOrigin.Begin);
-
-            return await StoreAsync(
-                GenerateStoreRequestUri(studyInstanceUid),
-                ConvertStreamToStreamContent(stream),
-                cancellationToken)
-                .ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebResponse<DicomDataset>> StoreAsync(
-            DicomFile dicomFile,
-            string studyInstanceUid,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(dicomFile, nameof(dicomFile));
-
-            await using MemoryStream stream = GetMemoryStream();
-            await dicomFile.SaveAsync(stream).ConfigureAwait(false);
-            stream.Seek(0, SeekOrigin.Begin);
-
-            using StreamContent content = ConvertStreamToStreamContent(stream);
-            return await StoreAsync(
-                GenerateStoreRequestUri(studyInstanceUid),
-                content,
-                cancellationToken)
-                .ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebResponse<DicomDataset>> StoreAsync(
-            HttpContent content,
-            CancellationToken cancellationToken)
-        {
-            return await StoreAsync(
-                new Uri("/" + _apiVersion + DicomWebConstants.StudiesUriString, UriKind.Relative),
-                content,
-                cancellationToken)
-                .ConfigureAwait(false);
-        }
-
-        private async Task<DicomWebResponse<DicomDataset>> StoreAsync(Uri requestUri, HttpContent content, CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(requestUri, nameof(requestUri));
-            EnsureArg.IsNotNull(content, nameof(content));
-
-            using var request = new HttpRequestMessage(HttpMethod.Post, requestUri);
-
-            request.Headers.Accept.Add(DicomWebConstants.MediaTypeApplicationDicomJson);
-
-            request.Content = content;
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(
-                response,
-                additionalFailureInspector: (statusCode, responseHeaders, contentHeaders, responseBody) =>
-                {
-                    // If store fails, we will get Conflict status code but the body will be a DicomDataset,
-                    // so we need to handle this case specially.
-                    if (statusCode == HttpStatusCode.Conflict)
-                    {
-                        throw new DicomWebException(
-                            statusCode,
-                            responseHeaders,
-                            contentHeaders,
-                            JsonConvert.DeserializeObject<DicomDataset>(responseBody, _jsonSerializerSettings));
-                    }
-
-                    return false;
-                })
-                .ConfigureAwait(false);
-
-            return new DicomWebResponse<DicomDataset>(
-                response,
-                async content =>
-                {
-                    string contentText = await content.ReadAsStringAsync().ConfigureAwait(false);
-                    return JsonConvert.DeserializeObject<DicomDataset>(contentText, _jsonSerializerSettings);
-                });
-        }
-        #endregion
-
-        #region Delete
-        public async Task<DicomWebResponse> DeleteStudyAsync(
-            string studyInstanceUid,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-
-            var requestUri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseStudyUriFormat, studyInstanceUid), UriKind.Relative);
-
-            return await DeleteAsync(requestUri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebResponse> DeleteSeriesAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-
-            var requestUri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseSeriesUriFormat, studyInstanceUid, seriesInstanceUid), UriKind.Relative);
-
-            return await DeleteAsync(requestUri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebResponse> DeleteInstanceAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string sopInstanceUid,
-            CancellationToken cancellationToken = default)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(sopInstanceUid, nameof(sopInstanceUid));
-
-            var requestUri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.BaseInstanceUriFormat, studyInstanceUid, seriesInstanceUid, sopInstanceUid), UriKind.Relative);
-
-            return await DeleteAsync(requestUri, cancellationToken).ConfigureAwait(false);
-        }
-
-        private async Task<DicomWebResponse> DeleteAsync(Uri requestUri, CancellationToken cancellationToken)
-        {
-            using var request = new HttpRequestMessage(HttpMethod.Delete, requestUri);
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebResponse(response);
-        }
-        #endregion
-
-        #region QIDO-RS
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QueryStudyAsync(
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            var uri = new Uri("/" + _apiVersion + DicomWebConstants.StudiesUriString + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QueryStudySeriesAsync(
-            string studyInstanceUid,
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-
-            var uri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.QueryStudySeriesUriFormat, studyInstanceUid) + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QueryStudyInstanceAsync(
-            string studyInstanceUid,
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-
-            var uri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.QueryStudyInstanceUriFormat, studyInstanceUid) + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QueryStudySeriesInstanceAsync(
-            string studyInstanceUid,
-            string seriesInstanceUid,
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(studyInstanceUid, nameof(studyInstanceUid));
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-
-            var uri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.QueryStudySeriesInstancesUriFormat, studyInstanceUid, seriesInstanceUid) + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QuerySeriesAsync(
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            var uri = new Uri("/" + _apiVersion + DicomWebConstants.SeriesUriString + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QuerySeriesInstanceAsync(
-            string seriesInstanceUid,
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(seriesInstanceUid, nameof(seriesInstanceUid));
-
-            var uri = new Uri("/" + _apiVersion + string.Format(DicomWebConstants.QuerySeriesInstancUriFormat, seriesInstanceUid) + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QueryInstancesAsync(
-            string queryString,
-            CancellationToken cancellationToken)
-        {
-            var uri = new Uri("/" + _apiVersion + DicomWebConstants.InstancesUriString + GetQueryParamUriString(queryString), UriKind.Relative);
-
-            return await QueryAsync(uri, cancellationToken).ConfigureAwait(false);
-        }
-
-        private async Task<DicomWebAsyncEnumerableResponse<DicomDataset>> QueryAsync(
-            Uri requestUri,
-            CancellationToken cancellationToken)
-        {
-            using var request = new HttpRequestMessage(HttpMethod.Get, requestUri);
-
-            request.Headers.Accept.Add(DicomWebConstants.MediaTypeApplicationDicomJson);
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebAsyncEnumerableResponse<DicomDataset>(
-                response,
-                DeserializeAsAsyncEnumerable<DicomDataset>(response.Content));
-        }
-        #endregion
-
-        #region ChangeFeed
-        public async Task<DicomWebAsyncEnumerableResponse<ChangeFeedEntry>> GetChangeFeed(string queryString, CancellationToken cancellationToken)
-        {
-            using var request = new HttpRequestMessage(
-                HttpMethod.Get,
-                new Uri($"/{_apiVersion}/changefeed{queryString}", UriKind.Relative));
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebAsyncEnumerableResponse<ChangeFeedEntry>(
-                response,
-                DeserializeAsAsyncEnumerable<ChangeFeedEntry>(response.Content));
-        }
-
-        public async Task<DicomWebResponse<ChangeFeedEntry>> GetChangeFeedLatest(string queryString, CancellationToken cancellationToken)
-        {
-            using var request = new HttpRequestMessage(
-                HttpMethod.Get,
-                new Uri($"/{_apiVersion}/changefeed/latest{queryString}", UriKind.Relative));
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebResponse<ChangeFeedEntry>(
-                response,
-                async content =>
-                {
-                    string contentText = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                    return JsonConvert.DeserializeObject<ChangeFeedEntry>(contentText, _jsonSerializerSettings);
-                });
-        }
-        #endregion
-
-        #region ExtendedQueryTag
-        public async Task<DicomWebResponse> AddExtendedQueryTagAsync(IEnumerable<AddExtendedQueryTagEntry> tagEntries, CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNull(tagEntries, nameof(tagEntries));
-            string jsonString = JsonConvert.SerializeObject(tagEntries);
-            var uri = new Uri($"/{_apiVersion}{DicomWebConstants.BaseExtendedQueryTagUri}", UriKind.Relative);
-            using var request = new HttpRequestMessage(HttpMethod.Post, uri);
-            {
-                request.Content = new StringContent(jsonString);
-                request.Content.Headers.ContentType = new System.Net.Http.Headers.MediaTypeHeaderValue(DicomWebConstants.ApplicationJsonMediaType);
-            }
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-            return new DicomWebResponse(response);
-        }
-
-        public async Task<DicomWebResponse> DeleteExtendedQueryTagAsync(string tagPath, CancellationToken cancellationToken)
-        {
-            EnsureArg.IsNotNullOrWhiteSpace(tagPath, nameof(tagPath));
-
-            var uri = new Uri($"/{_apiVersion}{DicomWebConstants.BaseExtendedQueryTagUri}/{tagPath}", UriKind.Relative);
-            using var request = new HttpRequestMessage(HttpMethod.Delete, uri);
-
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-
-            return new DicomWebResponse(response);
-        }
-
-        public async Task<DicomWebResponse<IEnumerable<GetExtendedQueryTagEntry>>> GetExtendedQueryTagsAsync(CancellationToken cancellationToken)
-        {
-            var uri = new Uri($"/{_apiVersion}{DicomWebConstants.BaseExtendedQueryTagUri}", UriKind.Relative);
-            using var request = new HttpRequestMessage(HttpMethod.Get, uri);
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-            return new DicomWebResponse<IEnumerable<GetExtendedQueryTagEntry>>(
-                 response,
-                 async content =>
-                 {
-                     string contentText = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                     return JsonConvert.DeserializeObject<IEnumerable<GetExtendedQueryTagEntry>>(contentText, _jsonSerializerSettings);
-                 });
-        }
-
-        public async Task<DicomWebResponse<GetExtendedQueryTagEntry>> GetExtendedQueryTagAsync(string tagPath, CancellationToken cancellationToken)
-        {
-            var uri = new Uri($"/{_apiVersion}{DicomWebConstants.BaseExtendedQueryTagUri}/{tagPath}", UriKind.Relative);
-            using var request = new HttpRequestMessage(HttpMethod.Get, uri);
-            HttpResponseMessage response = await HttpClient.SendAsync(request, cancellationToken)
-                .ConfigureAwait(false);
-            await EnsureSuccessStatusCodeAsync(response).ConfigureAwait(false);
-            return new DicomWebResponse<GetExtendedQueryTagEntry>(
-                 response,
-                 async content =>
-                 {
-                     string contentText = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                     return JsonConvert.DeserializeObject<GetExtendedQueryTagEntry>(contentText, _jsonSerializerSettings);
-                 });
-        }
-        #endregion
-
-        #region Helpers
         [SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope", Justification = "Callers will dispose of the StreamContent")]
         private static MultipartContent ConvertStreamsToMultipartContent(IEnumerable<Stream> streams)
         {
@@ -806,6 +193,5 @@ namespace Microsoft.Health.Dicom.Client
                 yield return item;
             }
         }
-        #endregion
     }
 }
