+++ src/Microsoft.Health.Dicom.Client/IDicomWebClient.Delete.cs
@@ -1,0 +1,18 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.Health.Dicom.Client
+{
+    public partial interface IDicomWebClient
+    {
+        Task<DicomWebResponse> DeleteExtendedQueryTagAsync(string tagPath, CancellationToken cancellationToken = default);
+        Task<DicomWebResponse> DeleteInstanceAsync(string studyInstanceUid, string seriesInstanceUid, string sopInstanceUid, CancellationToken cancellationToken = default);
+        Task<DicomWebResponse> DeleteSeriesAsync(string studyInstanceUid, string seriesInstanceUid, CancellationToken cancellationToken = default);
+        Task<DicomWebResponse> DeleteStudyAsync(string studyInstanceUid, CancellationToken cancellationToken = default);
+    }
+}
