+++ src/Microsoft.Health.Dicom.Client/IDicomWebClient.ExtendedQueryTag.cs
@@ -3,15 +3,16 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System;
-using System.IO;
-using System.Net.Http;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Health.Dicom.Client.Models;
 
 namespace Microsoft.Health.Dicom.Client
 {
     public partial interface IDicomWebClient
     {
-        Func<MemoryStream> GetMemoryStream { get; set; }
-        HttpClient HttpClient { get; }
+        Task<DicomWebResponse<OperationReference>> AddExtendedQueryTagAsync(IEnumerable<AddExtendedQueryTagEntry> tagEntries, CancellationToken cancellationToken = default);
+        Task<DicomWebResponse> DeleteExtendedQueryTagAsync(string tagPath, CancellationToken cancellationToken = default);
     }
 }
