+++ src/Microsoft.Health.Dicom.Client/Microsoft.Health.Dicom.Client.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <Description>Defines a RESTful client for interacting with DICOMweb APIs.</Description>
@@ -14,6 +14,16 @@
     <PackageReference Include="fo-dicom" Version="4.0.7" />
     <PackageReference Include="fo-dicom.Json" Version="4.0.7" NoWarn="NU1701" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="IDicomWebClient.*.cs">
+      <DependentUpon>IDicomWebClient.cs</DependentUpon>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="DicomWebClient.*.cs">
+      <DependentUpon>DicomWebClient.cs</DependentUpon>
+    </None>
   </ItemGroup>
 
 </Project>
